[feature] 세션 데이터 - html 연결
