+++ moodico/main/admin.py
@@ -1,3 +1,7 @@
 from django.contrib import admin
+from .models import RankedProduct, VotingSession
 
 # Register your models here.
+
+admin.site.register(RankedProduct)
+admin.site.register(VotingSession)
